requirements/production.txt
-r base.txt

dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
django-storages==1.11.1
boto==2.49.0
django-redis-cache==3.0.0
django-anymail==8.2
raven==6.10.0
# opbeat==3.6.1
djangorestframework-jwt==1.11.0
boto3

